requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
streamlit
python-dotenv
fpdf
python-docx
schedule
openpyxl
tensorflow>=2.12.0; extra == "autoencoder"
faker

